pyproject.toml
[project]
name = "epstein-dossier"
version = "0.1.0"
description = "Comprehensive search platform for Epstein DOJ files"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",

    # Search
    "meilisearch>=0.31.0",

    # Vector DB
    "chromadb>=0.4.22",

    # PDF Processing
    "pymupdf>=1.23.20",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0",

    # NLP
    "spacy>=3.7.4",

    # Numeric
    "numpy>=1.26.3",
    "scikit-learn>=1.4.0",

    # HTTP Client
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "beautifulsoup4>=4.12.3",

    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "tqdm>=4.66.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
faces = [
    "face-recognition>=1.3.0",
]
graph = [
    "neo4j>=5.17.0",
]
all = [
    "face-recognition>=1.3.0",
    "neo4j>=5.17.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[project.scripts]
epstein = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
